requirements.txt
accelerate==0.19.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argilla==1.7.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
comtypes==1.1.14
cryptography==40.0.2
dataclasses-json==0.5.7
Deprecated==1.2.13
diffusers==0.16.1
et-xmlfile==1.1.0
faiss-cpu==1.7.4
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.5.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
joblib==1.2.0
keyboard==0.13.5
langchain==0.0.167
llvmlite==0.40.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numba==0.57.0
numexpr==2.8.4
numpy==1.23.5
olefile==0.46
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==9.5.0
psutil==5.9.5
PyAudio==0.2.13
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
pypandoc==1.11
pypiwin32==223
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
python-pptx==0.6.21
pyttsx3==2.90
pytz==2023.3
pywin32==306
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
rfc3986==1.5.0
rich==13.0.1
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.13
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.3
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers @ git+https://github.com/huggingface/transformers@15f260a82f98788354d55cb2788e9f0b5131fb77
typer==0.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
unstructured==0.6.6
urllib3==2.0.2
wrapt==1.14.1
XlsxWriter==3.1.0
yarl==1.9.2
zipp==3.15.0
